Removed duplicaters and renamed move* to arrange*
